:bug: Do NOT rely on @denops-private from @denops
