+++ denops/@denops-private/service/worker/script.ts
@@ -10,7 +10,7 @@ import {
   WorkerWriter,
 } from "../deps.ts";
 import { Denops, Meta } from "../../../@denops/denops.ts";
-import { DenopsImpl } from "../../../@denops-private/denops.ts";
+import { DenopsImpl } from "../../../@denops/denops.ts";
 
 // deno-lint-ignore no-explicit-any
 const worker = self as any as Worker;
